[RUMF-828] resolve tools/utils cyclic dependencies issues
Reduce tools/utils dependencies by moving stuff that requires internal
monitoring in a separate file (events-related functions). This allows
using tools/utils in internalMonitoring
